+++ include/lomse_interactor.h
@@ -331,6 +331,19 @@ public:
     */
     void switch_task(int taskType);
 
+    /** Define the duration for one beat, for metronome and for methods that use
+        measure/beat parameters to define a location. This value is shared by all
+        scores contained in the document and can be changed at any time.
+        Changes while the score is being played back are ignored until playback finishes.
+        @param beatType A value from enum #EBeatDuration.
+        @param duration The duration (in Lomse Time Units) for one beat. You can use
+            a value from enum ENoteDuration casted to double. This parameter is
+            required only when value for parameter `beatType` is `k_beat_specified`.
+            For all other values, if a non-zero value is specified, the value
+            will be used for the beat duration in scores without time signature.
+    */
+    void define_beat(int beatType, TimeUnits duration=0.0);
+
     //@}    //operating modes